# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "This is the API documentation for My API.",
        "termsOfService": "https://dev.test/terms",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://dev.test",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "description": "Returns a list of all products.",
                "operationId": "getProducts",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "List of products"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tests": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Get all tests",
                "operationId": "5f539f69bb1d910182eb35136c5baa3a",
                "responses": {
                    "200": {
                        "description": "List of tests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "List of tests retrieved successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Test"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tests"
                ],
                "summary": "Create a new test",
                "operationId": "7728a2f3dd87105d6d617df9a3f231d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Test Name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Test description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Test created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Test"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/tests/{id}": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Get a specific test",
                "operationId": "7e3d8428f4df82c6d4ee7bd1d4d2128c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the test",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Test retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Test"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Test not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tests"
                ],
                "summary": "Update a specific test",
                "operationId": "ca1490751234c723e0a4a708afe16dd6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the test",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Test Name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Test updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Test"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Test not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tests"
                ],
                "summary": "Delete a specific test",
                "operationId": "92f76a68796679554c71a4659f62b296",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the test",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Test deleted successfully."
                    },
                    "404": {
                        "description": "Test not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "title": "Product",
                "required": [
                    "id",
                    "name",
                    "price"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "iPhone 13"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 999.99
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "Test": {
                "title": "Test",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sample Test"
                    },
                    "description": {
                        "type": "string",
                        "example": "This is a test record."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Tests",
            "description": "API Endpoints for managing tests"
        },
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}